+++ CHANGELOG.md
@@ -37,7 +37,7 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 - UM-837 - Replay will now display high water mark statistics per allocator.
 
-- Added a benchmark that measures the performance of FixedPool across various allocation sizes.
+- Added a benchmark that measures the performance of FixedPool across two allocation sizes.
 
 ### Changed